backend/requirements.txt
fastapi
uvicorn
openai
python-multipart
pillow 
dotenv
asyncpg
SQLAlchemy>=2.0
python-dotenv
greenlet
asyncpg 
numpy
requests